task_manager/tasks/serializers.py
from rest_framework import serializers
from .models import Task, TaskPhoto

class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = '__all__'

class TaskPhotoSerializer(serializers.ModelSerializer):
    class Meta:
        model = TaskPhoto
        fields = '__all__'
